// src/types/color.ts
export type BasicColorGray =
	| "gray-000"
	| "gray-050"
	| "gray-100"
	| "gray-200"
	| "gray-300"
	| "gray-400"
	| "gray-500"
	| "gray-600"
	| "gray-700"
	| "gray-800"
	| "gray-900"
	| "gray-950";

export type BasicColorRed =
	| "red-050"
	| "red-100"
	| "red-200"
	| "red-300"
	| "red-400"
	| "red-500"
	| "red-600"
	| "red-700"
	| "red-800"
	| "red-900";

export type BasicColorPink =
	| "pink-050"
	| "pink-100"
	| "pink-200"
	| "pink-300"
	| "pink-400"
	| "pink-500"
	| "pink-600"
	| "pink-700"
	| "pink-800"
	| "pink-900";

export type BasicColorGrape =
	| "grape-050"
	| "grape-100"
	| "grape-200"
	| "grape-300"
	| "grape-400"
	| "grape-500"
	| "grape-600"
	| "grape-700"
	| "grape-800"
	| "grape-900";

export type BasicColorViolet =
	| "violet-050"
	| "violet-100"
	| "violet-200"
	| "violet-300"
	| "violet-400"
	| "violet-500"
	| "violet-600"
	| "violet-700"
	| "violet-800"
	| "violet-900";

export type BasicColorBlue =
	| "blue-050"
	| "blue-100"
	| "blue-200"
	| "blue-300"
	| "blue-400"
	| "blue-500"
	| "blue-600"
	| "blue-700"
	| "blue-800"
	| "blue-900";

export type BasicColorCyan =
	| "cyan-050"
	| "cyan-100"
	| "cyan-200"
	| "cyan-300"
	| "cyan-400"
	| "cyan-500"
	| "cyan-600"
	| "cyan-700"
	| "cyan-800"
	| "cyan-900";

export type BasicColorGreen =
	| "green-050"
	| "green-100"
	| "green-200"
	| "green-300"
	| "green-400"
	| "green-500"
	| "green-600"
	| "green-700"
	| "green-800"
	| "green-900";

export type BasicColorLime =
	| "lime-050"
	| "lime-100"
	| "lime-200"
	| "lime-300"
	| "lime-400"
	| "lime-500"
	| "lime-600"
	| "lime-700"
	| "lime-800"
	| "lime-900";

export type BasicColorYellow =
	| "yellow-050"
	| "yellow-100"
	| "yellow-200"
	| "yellow-300"
	| "yellow-400"
	| "yellow-500"
	| "yellow-600"
	| "yellow-700"
	| "yellow-800"
	| "yellow-900";

export type BasicColorOrange =
	| "orange-050"
	| "orange-100"
	| "orange-200"
	| "orange-300"
	| "orange-400"
	| "orange-500"
	| "orange-600"
	| "orange-700"
	| "orange-800"
	| "orange-900";

export type BasicColor =
	| BasicColorGray
	| BasicColorRed
	| BasicColorPink
	| BasicColorGrape
	| BasicColorViolet
	| BasicColorBlue
	| BasicColorCyan
	| BasicColorGreen
	| BasicColorLime
	| BasicColorYellow
	| BasicColorOrange;

export type SemanticColorTheme =
	| "primary"
	| "secondary"
	| "success"
	| "warning"
	| "danger"
	| "contrast"
	| "hint";

export type SemanticColorThemeTransparent =
	| "primary-transparent-8"
	| "secondary-transparent-8"
	| "success-transparent-8"
	| "warning-transparent-8"
	| "danger-transparent-8"
	| "hint-transparent-8"
	| "contrast-transparent-8"
	| "primary-transparent-16"
	| "secondary-transparent-16"
	| "success-transparent-16"
	| "warning-transparent-16"
	| "danger-transparent-16"
	| "hint-transparent-16"
	| "contrast-transparent-16"
	| "primary-transparent-24"
	| "secondary-transparent-24"
	| "success-transparent-24"
	| "warning-transparent-24"
	| "danger-transparent-24"
	| "hint-transparent-24"
	| "contrast-transparent-24"
	| "primary-transparent-32"
	| "secondary-transparent-32"
	| "success-transparent-32"
	| "warning-transparent-32"
	| "danger-transparent-32"
	| "hint-transparent-32"
	| "contrast-transparent-32";

export type SemanticColorStatus =
	| "primary-hover"
	| "secondary-hover"
	| "success-hover"
	| "warning-hover"
	| "danger-hover"
	| "hint-hover"
	| "contrast-hover"
	| "primary-active"
	| "secondary-active"
	| "success-active"
	| "warning-active"
	| "danger-active"
	| "hint-active"
	| "contrast-active";

export type SemanticColorText =
	| "text-primary"
	| "text-secondary"
	| "text-success"
	| "text-warning"
	| "text-danger"
	| "text-hint"
	| "text-contrast"
	| "text-primary-alternative"
	| "text-secondary-alternative"
	| "text-success-alternative"
	| "text-warning-alternative"
	| "text-danger-alternative"
	| "text-hint-alternative"
	| "text-contrast-alternative"
	| "text-alternative"
	| "text-normal"
	| "text-white"
	| "text-exception";

export type SemanticColorBackground =
	| "background-normal"
	| "background-alternative-01"
	| "background-alternative-02";

export type SemanticColorBorder = "border-color" | "border-hover";